fix: fixed inconsistent deleting between chefs and restaurants

+++ frontend/src/app/components/restaurant-list-component/restaurant-list-component.ts
@@ -48,9 +48,13 @@ export class RestaurantListComponent {
 
   deleteRestaurant(id: number): void {
     //const headers = { Authorization: 'Basic ' + btoa('user:pass') };
+    if (!confirm('Are you sure you want to delete this restaurant?')) return;
     this.http.delete(`http://localhost:8081/api/restaurant/${id}`)
       .subscribe({
-        next: () => this.fetchRestaurants(),
+        next: () => {
+
+            this.restaurants = this.restaurants.filter(c => c.id !== id);
+        },
         error: (err) => this.error = 'Failed to delete restaurant.'
       });
   }
